Changed duration to integer
Added minute and second to Log and Functions VC

+++ Think+/Database/Session.swift
@@ -13,16 +13,16 @@ class Session {
     let id: Int64?
     var sessionDate: String
     var sessionTime: String
-    var duration: Double
+    var duration: Int
     
     init(id: Int64) {
         self.id = id
         sessionDate = ""
         sessionTime = ""
-        duration = 0.0
+        duration = 0
     }
     
-    init(id: Int64, sessionDate: String, sessionTime: String, duration: Double) {
+    init(id: Int64, sessionDate: String, sessionTime: String, duration: Int) {
         self.id = id
         self.sessionDate = sessionDate
         self.sessionTime = sessionTime
